refactor(SectioningContent): client境界が必要なモジュールをclient/配下に分離
SectioningContent.tsxからuseContextをSectioningFragmentへ切り出し、
'use client'を外してServer Componentとして利用できるようにした。

client/配下は境界の必要性を表す。componentsには'use client'を設定し、
hooksには設定しない（利用側のコンポーネントが境界を持つ方針のため）。

分離にあたりバレルの構成が重要になる。rollupはpreserveModulesでも
バレルを平坦化して実体へ直リンクするため、バレルが依存するモジュールの
外部依存が呼び出し側へ副作用importとして転記される。
componentsとhooksを別バレルに分けることで、SectioningContent.jsへの
`import 'styled-components'`の混入を防いでいる。

検証:
- ビルド出力をreact-server条件で評価し、SectioningContent.jsがOKになることを確認
- 公開バレルからSectionのみimportした場合、styled-componentsが
  ツリーシェイクで落ちることをesbuildで確認
- DOM出力は13パターンでmasterと完全一致
