+++ packages/smarthr-ui/src/components/SectioningContent/client/components/SectioningFragment.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import { type FC, type PropsWithChildren, useContext } from 'react'
+
+import { LevelContext } from './LevelContext'
+
+export const SectioningFragment: FC<PropsWithChildren<{ baseLevel?: number }>> = ({
+  children,
+  baseLevel,
+}) => {
+  const level = useContext(LevelContext)
+
+  return <LevelContext.Provider value={baseLevel || level + 1}>{children}</LevelContext.Provider>
+}
